Added functionality to the parties package. Need vmarket Implementation.
I pushed several attributes up the hierarchy. I added some functionality
to the Seller and Buyer interface. I added a List<Party> to the Manager
class, which will allow the Manager class to "manage" those Parties. I
added a List<Commodity> to each class that implements Buyer to hold a
reference to the Commodities he seeks to purchase. 

For actual implementation of many of the classes in the vmarket/parties
package, there is functionality required that is absent in the vmarket
package. To be specific, the following is required:

 1. A central data structure that holds all of the Commodities available
for purchase.
 2. The central data structure needs to be able to hold the following
information, either as a part of the Commodity or the data structure
itself:
 	a. The Commodity to be sold.
 	b. The amount of the Commodity.
 	c. The Price of the Commodity.
 	d. A Reference(Party) to the owner of the Commodity. 
 3. Methods to get Commodities, add Commodities, or Remove Commodities
from the aforementioned central data structure, and for the attributes
listed in #2.